requirements.txt
llama-cpp-python>=0.2.56

chromadb>=0.4.22

sentence-transformers>=2.7.0
nomic>=2.0.0

rank-bm25>=0.2.2

pdfplumber>=0.11.0
beautifulsoup4>=4.12.0

gradio>=4.20.0

numpy>=1.26.0
pathlib>=1.0.1

einops>=0.7.0